default imports -> imports; renamed file of useAuthStatus hook
