+++ src/hooks/index.ts
@@ -1,9 +1,8 @@
-import useMap from './use-map.ts';
-import useCityLocation from './use-city-location.ts';
-import useMarkers from './use-markers.ts';
-import useSelect from './use-select.ts';
+import { useMap } from './use-map.ts';
+import { useCityLocation } from './use-city-location.ts';
+import { useMarkers } from './use-markers.ts';
 import { useBooleanHelpers } from './use-boolean-helpers.ts';
-import { useAuthStatus } from './routing.ts';
+import { useAuthStatus } from './use-auth-status.ts';
 import { useFavoriteOffers } from './use-favorite-offers.ts';
 import { useNearbyOffers } from './use-nearby-offers.ts';
 import { useDetailOffer } from './use-detail-offer.ts';
@@ -12,7 +11,6 @@ export {
   useMap,
   useCityLocation,
   useMarkers,
-  useSelect,
   useBooleanHelpers,
   useAuthStatus,
   useFavoriteOffers,
