+++ src/hooks/use-auth-status.ts
